Add files via upload
Figure1. Time Series plot of exploit dynamics (Microsoft Products)

In Figure.1, the difference between exploit date and published date in days is placed on the vertical axis, and CVE ids (sorted in ascending order) are located on the x-axis. As is shown, the trend is almost steady which means that most of the exploits were developed immediately after the vulnerability had been published. Moreover, part of the mentioned exploits were developed even before the published date of relative vulnerabilities.