Add files via upload
Figure.2 Cumulative distribution function of 'days2exploit'  (Microsoft Products)
In Figure.2, the cumulative probability distribution of exploit dynamics is presented with the number of days to exploit on x-axis and cumulative probability on y-axis. The figure shows that approximately 20% of vulnerabilities were exploited before they were published during the period of study.